test: refactor 2 tests from Enzyme to RTL (#15601)
